pyproject.toml
[project]
name = "pasci-api"
version = "0.1.0"
description = "API for PASCI Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.121.1",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.4",
    "sqlmodel>=0.0.27",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.2",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.13",
    "python-magic>=0.4.27",
    "pydantic-settings>=2.12.0",
    "asyncpg>=0.31.0",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]
